Add FieldManager.Home
