+++ Maple2.Server.Game/Manager/Field/FieldManager.cs
@@ -71,7 +71,6 @@ public sealed partial class FieldManager : IDisposable {
     public ItemDropManager ItemDrop { get; }
 
     public int MapId => Metadata.Id;
-    public readonly long OwnerId;
     public readonly int InstanceId;
     public FieldInstance FieldInstance = FieldInstance.Default;
     public readonly AiManager Ai;
